Update the evaluations base, found some stuff that would help on the data condition evaluation result for the processed data condition replacements
