+++ src/sentry/workflow_engine/processors/evaluations/__init__.py
@@ -1,6 +1,8 @@
 __all__ = [
     "DataConditionEvaluation",
     "DataConditionEvaluationException",
+    "TriggerResult",
 ]
 
 from .condition import DataConditionEvaluation, DataConditionEvaluationException
+from .trigger_result import TriggerResult
