+++ src/sentry/workflow_engine/processors/evaluations/base.py
@@ -1,16 +1,11 @@
-from abc import ABC
+from dataclasses import dataclass
 
 
-class BaseWorkflowEngineEvaluation(ABC):
+@dataclass(frozen=True, kw_only=True)
+class BaseWorkflowEngineEvaluation[R, E]:
     """
     This is a shared base class for all Evaluation classes.
-
-    TODO
-    - Once we have a few of these, normalize them here
-
-    Current candidates:
-    - value: Any -- the value used in the evaluation
-    - result: Any -- the value returned as the result
     """
 
-    pass
+    result: R | None
+    error: E | None
